@@ -11,17 +11,17 @@
   "type": "cakephp-plugin",
   "license": ["MIT"],
   "require": {
-    "php": ">=7.2",
+    "php": ">=8.1",
     "ext-json": "*",
-    "cakephp/core": "^4.3",
-    "cakephp/orm": "^4.0",
-    "cakephp/authentication": "^2.0"
+    "cakephp/core": "^5.0",
+    "cakephp/orm": "^5.0",
+    "cakephp/authentication": "^3.0"
   },
   "require-dev": {
-    "cakephp/cakephp": "^4.3",
-    "cakephp/cakephp-codesniffer": "^4.6",
-    "cakephp/migrations": "^3.3",
-    "phpunit/phpunit": "^8.5 || ^9.3"
+    "cakephp/cakephp": "^5.0",
+    "cakephp/cakephp-codesniffer": "^5.1",
+    "cakephp/migrations": "^4.0",
+    "phpunit/phpunit": "^10.1"
   },
   "autoload": {
     "psr-4": {
