@@ -28,7 +28,7 @@ class RememberMeToken extends Entity
      *
      * @var array
      */
-    protected $_accessible = [
+    protected array $_accessible = [
         '*' => true,
         'id' => false,
     ];
@@ -38,7 +38,7 @@ class RememberMeToken extends Entity
      *
      * @var array
      */
-    protected $_hidden = [
+    protected array $_hidden = [
         'token',
     ];
 }
