@@ -15,7 +15,7 @@ class AuthUsersFixture extends TestFixture
      *
      * @var array
      */
-    public $records = [
+    public array $records = [
         ['id' => 1, 'username' => 'foo', 'password' => 'not use'],
         ['id' => 2, 'username' => 'bar', 'password' => 'not use'],
         ['id' => 3, 'username' => 'boo', 'password' => 'not use'],
