@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace RememberMe\Authenticator;
 
+use ArrayAccess;
 use Cake\Utility\Security;
 use InvalidArgumentException;
 
@@ -17,6 +18,7 @@ trait EncryptCookieTrait
      * @param string $cookie from request
      * @return array ['username' => ..., 'series' => ..., 'token' => ...]
      * @throws \InvalidArgumentException
+     * @throws \JsonException
      */
     public static function decodeCookie(string $cookie): array
     {
@@ -25,7 +27,7 @@ trait EncryptCookieTrait
             throw new InvalidArgumentException('Can\'t decrypt cookie.');
         }
 
-        return json_decode($decryptedValue, true);
+        return json_decode($decryptedValue, true, 512, JSON_THROW_ON_ERROR);
     }
 
     /**
@@ -35,12 +37,13 @@ trait EncryptCookieTrait
      * @param string $series series string
      * @param string $token login token
      * @return string
+     * @throws \JsonException
      */
     public static function encryptToken(string $username, string $series, string $token): string
     {
         return base64_encode(
             Security::encrypt(
-                json_encode(compact('username', 'series', 'token')),
+                json_encode(compact('username', 'series', 'token'), JSON_THROW_ON_ERROR),
                 Security::getSalt()
             )
         );
@@ -52,7 +55,7 @@ trait EncryptCookieTrait
      * @param \ArrayAccess|array $identity logged in user info
      * @return string
      */
-    protected static function _generateToken($identity): string
+    protected static function _generateToken(ArrayAccess|array $identity): string
     {
         $prefix = bin2hex(Security::randomBytes(16));
 
